implement protobuf writer, refactor msgpack serialisation to reuse code
