+++ dd-trace-core/src/main/java/datadog/trace/core/serialization/ValueWriter.java
@@ -1,0 +1,6 @@
+package datadog.trace.core.serialization;
+
+// TODO @FunctionalInterface
+public interface ValueWriter<T> {
+  void write(T value, Writable writable, EncodingCache encodingCache);
+}
